bug fix
re arranging everything
